moved readOtgStdin envSubstOrDefault to create.go
