+++ cmd/flow.go
@@ -23,10 +23,7 @@ package cmd
 
 import (
 	"fmt"
-	"io"
-	"os"
 
-	"github.com/drone/envsubst"
 	"github.com/open-traffic-generator/snappi/gosnappi"
 	"github.com/spf13/cobra"
 )
@@ -303,32 +300,4 @@ func newFlow(config gosnappi.Config) {
 		log.Fatal(err)
 	}
 	fmt.Print(otgYaml)
-}
-
-func readOtgStdin(api gosnappi.GosnappiApi) gosnappi.Config {
-	otgbytes, err := io.ReadAll(os.Stdin)
-	if err != nil {
-		log.Fatal(err)
-	}
-	otg := string(otgbytes)
-
-	config := api.NewConfig()
-	err = config.FromYaml(otg) // Thus YAML is assumed by default, and as a superset of JSON, it works for JSON format too
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return config
-}
-
-// Substitute e with env variable of such name, if it is not empty, otherwise use default vaule d
-func envSubstOrDefault(e string, d string) string {
-	s, err := envsubst.EvalEnv(e)
-	if err != nil {
-		log.Fatal(err)
-	}
-	if s == "" {
-		s = d
-	}
-	return s
 }
